requirements.txt
alabaster==0.7.12
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
gunicorn==20.0.4
idna==2.9
imagesize==1.2.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.3
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.6
pystache==0.5.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
smartypants==2.0.1
snowballstemmer==2.0.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
